+++ src/exceptions/index.ts
@@ -1,0 +1,7 @@
+export * from './validation';
+export * from './grpc/grpc-error';
+export * from './grpc/not-found';
+export * from './http/bad-request';
+export * from './http/http-error';
+export * from './http/not-found';
+export * from './http/unauthorized';