+++ src/exceptions/validation.ts
@@ -1,0 +1,8 @@
+import Joi from 'joi';
+import { CodedError } from '../types/exception';
+
+export class ValidationError extends CodedError {
+  constructor(details: Joi.ValidationErrorItem[]) {
+    super('VALIDATION_FAILED', 'Invalid request data', details);
+  }
+}