+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from './types/http.type';
-export * from './types/module.type';
+export * from './types';
+export * from './exceptions';