+++ src/types/entity.ts
@@ -1,0 +1,8 @@
+export abstract class Entity<T> {
+  constructor(props: Partial<T>) {
+    this.validate(props);
+    Object.assign(this, props);
+  }
+
+  protected abstract validate(props: Partial<T>): void;
+}